Expand the compendium compilation flow
Schema generation:
This adds a Formatter to the schema generator to automatically add a
regex pattern to the `_id` field.
It also removes the watch task for the schema generation task, since it
didn't work with longer running threads. (cached, like template task)

Compendium compilation:
This adds a compendium compile task, as well as a compendium entry for
weapons to the system.json.