Enable exactOptionalPropertyTypes for TS sources
This required a massive restructuring of the sources. The tsconfig.json
used for the main sources now lives with the typescript files directly.
This was necessary, because some dependencies used in the gulp files do
not support exactOptionalPropertyTypes. The main culprit here is AJV.
Further some aspects of how tsd works made it necessary to have tsd use
that same moved tsconfig.json file, so the test sources now also live in
src.